# Optimise getItemDataFromClassName and map hash throttle