Optimise some functions that are called repeatedly when loading map objects.

I didn't really want to deep dive into more save file loading optimisation at the moment as there is quite a lot of logic in there, so these are just a couple of 'easy wins' at the moment.

Before Changes:
addMapLayers: 2634.0009765625 ms
addMapLayers: 2583.52880859375 ms


After Changes:
addMapLayers: 1786.555908203125 ms
addMapLayers: 1785.843017578125 ms